From Twitter: make some variables 64 bit
Summary: Part of patch for making per category counters. Part 2
coming. suitably edited by groys

Test Plan:
run testsuite

+++ src/scribe_server.h
@@ -79,7 +79,7 @@ class scribeHandler : virtual public scribe::thrift::scribeIf,
   time_t lastMsgTime;
   unsigned long numMsgLastSecond;
   unsigned long maxMsgPerSecond;
-  unsigned long maxQueueSize;
+  unsigned long long maxQueueSize;
   bool newThreadPerCategory;
 
   /* mutex to syncronize access to scribeHandler.
